+++ plugin/src/main/java/net/codingarea/challenges/plugin/utils/bukkit/nms/implementations/v1_13/PlayerConnection_1_13.java
@@ -1,4 +1,4 @@
-package net.codingarea.challenges.plugin.utils.bukkit.nms.implementations.v_1_13;
+package net.codingarea.challenges.plugin.utils.bukkit.nms.implementations.v1_13;
 
 import net.codingarea.challenges.plugin.Challenges;
 import net.codingarea.challenges.plugin.utils.bukkit.nms.NMSUtils;
